broken down category store into categoryFilter and userCategories
